Compiles landing page into it's own component
